requirements.txt
discord.py
psutil
python-dotenv
requests
geoip2
aiohttp
factorio-rcon-py